More realistic JS bench
For the intended use-case, which is as the writer for a serde library,
the actual byte slices being passed will typically be very small, a
couple bytes at a time, with the occasional burst of larger ones for
strings. This chunks the input random bytes up, to better simulate that
behaviour vs. a flat write loop.